+++ tests/serialization/test_storage.py
@@ -9,16 +9,20 @@ from prefect.serialization.storage import (
     BaseStorageSchema,
     BytesSchema,
     DockerSchema,
+    GCSSchema,
     LocalSchema,
     MemorySchema,
+    S3Schema,
 )
 
 
-@pytest.mark.parametrize("cls", storage.Storage.__subclasses__())
-def test_serialization_on_all_subclasses(cls):
-    serialized = cls().serialize()
-    assert serialized
-    assert serialized["__version__"] == prefect.__version__
+def test_all_storage_subclasses_have_schemas():
+    "Test that ensures we don't forget to include a Schema for every subclass we implement"
+
+    subclasses = set(c.__name__ for c in storage.Storage.__subclasses__())
+    subclasses.add(storage.Storage.__name__)  # add base storage, not a subclass
+    schemas = set(prefect.serialization.storage.StorageSchema().type_schemas.keys())
+    assert subclasses == schemas
 
 
 def test_docker_empty_serialize():
@@ -82,6 +86,73 @@ def test_docker_serialize_with_flows():
     assert f.name in deserialized
 
 
+def test_s3_empty_serialize():
+    s3 = storage.S3(bucket="bucket")
+    serialized = S3Schema().dump(s3)
+
+    assert serialized
+    assert serialized["__version__"] == prefect.__version__
+    assert serialized["bucket"]
+    assert not serialized["key"]
+
+
+def test_s3_full_serialize():
+    s3 = storage.S3(
+        aws_access_key_id="id",
+        aws_secret_access_key="secret",
+        aws_session_token="session",
+        bucket="bucket",
+        key="key",
+    )
+    serialized = S3Schema().dump(s3)
+
+    assert serialized
+    assert serialized["__version__"] == prefect.__version__
+    assert serialized["bucket"] == "bucket"
+    assert serialized["key"] == "key"
+
+
+def test_s3_aws_creds_not_serialized():
+    s3 = storage.S3(
+        aws_access_key_id="id",
+        aws_secret_access_key="secret",
+        aws_session_token="session",
+        bucket="bucket",
+        key="key",
+    )
+    serialized = S3Schema().dump(s3)
+
+    assert serialized
+    assert serialized["__version__"] == prefect.__version__
+    assert serialized["bucket"] == "bucket"
+    assert serialized["key"] == "key"
+    assert serialized.get("aws_access_key_id") is None
+    assert serialized.get("aws_secret_access_key") is None
+    assert serialized.get("aws_session_token") is None
+
+
+def test_s3_serialize_with_flows():
+    s3 = storage.S3(
+        aws_access_key_id="id",
+        aws_secret_access_key="secret",
+        aws_session_token="session",
+        bucket="bucket",
+        key="key",
+    )
+    f = prefect.Flow("test")
+    s3.flows["test"] = "key"
+    serialized = S3Schema().dump(s3)
+
+    assert serialized
+    assert serialized["__version__"] == prefect.__version__
+    assert serialized["bucket"] == "bucket"
+    assert serialized["key"] == "key"
+    assert serialized["flows"] == {"test": "key"}
+
+    deserialized = S3Schema().load(serialized)
+    assert f.name in deserialized
+
+
 def test_bytes_empty_serialize():
     b = storage.Bytes()
     serialized = BytesSchema().dump(b)
@@ -123,3 +194,41 @@ def test_local_roundtrip():
         runner = deserialized.get_flow(flow_loc)
 
     assert runner.run().is_successful()
+
+
+def test_gcs_empty_serialize():
+    gcs = storage.GCS(bucket="bucket")
+    serialized = GCSSchema().dump(gcs)
+
+    assert serialized
+    assert serialized["__version__"] == prefect.__version__
+    assert serialized["bucket"]
+    assert not serialized["key"]
+
+
+def test_gcs_full_serialize():
+    gcs = storage.GCS(bucket="bucket", key="key", project="project",)
+    serialized = GCSSchema().dump(gcs)
+
+    assert serialized
+    assert serialized["__version__"] == prefect.__version__
+    assert serialized["bucket"] == "bucket"
+    assert serialized["key"] == "key"
+    assert serialized["project"] == "project"
+
+
+def test_gcs_serialize_with_flows():
+    gcs = storage.GCS(project="project", bucket="bucket", key="key",)
+    f = prefect.Flow("test")
+    gcs.flows["test"] = "key"
+    serialized = GCSSchema().dump(gcs)
+
+    assert serialized
+    assert serialized["__version__"] == prefect.__version__
+    assert serialized["bucket"] == "bucket"
+    assert serialized["key"] == "key"
+    assert serialized["project"] == "project"
+    assert serialized["flows"] == {"test": "key"}
+
+    deserialized = GCSSchema().load(serialized)
+    assert f.name in deserialized